ai-sso-system/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.2
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-dotenv==1.0.0
redis==5.0.1
httpx==0.25.2
python-multipart==0.0.5
aiohttp==3.9.1
celery==5.3.6  # For background tasks like email/sms
flower==2.0.1  # Optional for Celery monitoring